+++ .gitignore
@@ -1,2 +1,5 @@
 MyInvestor/MyInvestor.xcodeproj/project.xcworkspace/xcuserdata/*.xcuserdatad/UserInterfaceState.xcuserstate
 .DS_Store
+MyInvestor/.DS_Store
+*.xcuserstate
+MyInvestor/MyInvestor.xcodeproj/project.pbxproj
