--- MyInvestor/MyInvestor/Sources/ClientZone/Sources/Quote/Helpers/SKBAppUserBalance.swift
@@ -1,13 +1,0 @@
-//
-//  SKBAppUserBalance.swift
-//  MyInvestor
-//
-//  Created by Максим Скориков on 09.12.2025.
-//
-
-import Foundation
-
-struct UserBalance: Identifiable, Equatable {
-    let id = UUID()
-    let balance: Double
-}
