+++ MyInvestor/MyInvestor/Sources/ClientZone/Sources/Quote/Helpers/UserBalance+formatted.swift
@@ -1,21 +1,24 @@
 //
-//  SKBAppBalanceProperties.swift
+//  SKBAppUserBalance.swift
 //  MyInvestor
 //
 //  Created by Максим Скориков on 09.12.2025.
 //
 
-import SwiftUI
+import Foundation
+
+struct UserBalance: Identifiable, Equatable {
+    
+    let id = UUID()
+    let balance: Double
+}
 
 extension UserBalance {
     
     var formattedBalance: String {
-        
         let formatter = NumberFormatter()
         formatter.numberStyle = .currency
         formatter.maximumFractionDigits = 3
         return formatter.string(from: NSNumber(value: balance)) ?? "$\(balance)"
-        
     }
-    
 }
